[config] Fix legacy configs (#35)
## Summary

This fixes legacy configs that specify web-like services without using the `services` field. For example, using the jetpack SDK you can define some cronjobs in your code and publish a docker image. This doesn't need replicas or a deployment but does need the `image` field to be set in the app helm chart (because the register SDK job uses this field).

## How was it tested?

Tested `launchpad up` on a few projects:

* Project with legacy config that uses jetpack SDK with cronjobs
* Project with new config that publishes image. (with and without web type) 
* Project with new config that does not publish image.

## Is this change backwards-compatible?

Yes