Fix launchpad env set environment variable (#37)
## Summary
Right now, `launchpad env set` throws error:
```
jetconfig (launchpad.yaml) was not found.
```

This is because `launchpad env` commands assume that your project is in
the current directory, whereas `launchpad up` and other commands take
`path` as an argument. In `env` commands, the args are not related to
path, but they are treated as path, which is the wrong behavior. This PR
fixes that.

## How was it tested?
launchpad env set

## Is this change backwards-compatible?
Yes